+++ src/biz.dfch.CS.Appclusive.UI/Controllers/GatesController.cs
@@ -31,7 +31,7 @@ namespace biz.dfch.CS.Appclusive.UI.Controllers
         #region Gate
 
         // GET: Gates/Details/5
-        public ActionResult Details(long id, int rId = 0, string rAction = null, string rController = null)
+        public ActionResult Details(long id, long rId = 0, string rAction = null, string rController = null)
         {
             ViewBag.ReturnId = rId;
             ViewBag.ReturnAction = rAction;
