+++ src/biz.dfch.CS.Appclusive.UI/Controllers/JobsController.cs
@@ -29,7 +29,7 @@ namespace biz.dfch.CS.Appclusive.UI.Controllers
         protected override DataServiceQuery<Api.Core.Job> BaseQuery { get { return CoreRepository.Jobs; } }
 
         // GET: Jobs/Details/5
-        public ActionResult Details(long id, int rId = 0, string rAction = null, string rController = null)
+        public ActionResult Details(long id, long rId = 0, string rAction = null, string rController = null)
         {
             ViewBag.ReturnId = rId;
             ViewBag.ReturnAction = rAction;
